Append "**--gradient_accumulation_steps 10**" to the command line to enable the gradient accumulation
With an addition copy of gradient accumulation buffer, it can only fit batch_size of 5 and 9 for fp32 and fp16.

On 4 GPU Machine, batch_size=5, fp32 training
Single_GPU baseline: 9.89603 Ex/sec (with weight update step every batch)
With No gradient_accumulation, throughput  = 30.36 Ex/sec, scaling factor = 3.06
With gradient_accumulation_steps=10, throughput = 39.4464 Ex/sec , scaling factor = 3.98
With gradient_accumulation_steps=50, throughput = 41.1628 Ex/sec , scaling factor = 4.16 (it's large than 4, coz we do much less weight update step)

On 4 GPU Machine, batch_size=9, fp16 training
Single_GPU baseline 30.3243 Ex/sec (with weight update step every batch)
With No gradient_accumulation, throughput  = 79.0 Ex/sec, scaling factor = 2.60
With gradient_accumulation_steps=10, throughput = 111.95 Ex/sec , scaling factor = 3.69
With gradient_accumulation_steps=50, throughput = 123.13 Ex/sec , scaling factor =4.06 (it's large than 4, coz we do much less weight update step)

Related work items: #3887

+++ onnxruntime/test/training/runner/training_util.h
@@ -67,7 +67,7 @@ class DataSet {
 
 class RandomDataSet : public DataSet {
  public:
-  explicit RandomDataSet(int num_samples,
+  explicit RandomDataSet(size_t num_samples,
                          const std::vector<std::string>& tensor_names,
                          const std::vector<TensorShape> tensor_shapes,
                          const std::vector<onnx::TensorProto_DataType> tensor_types)
@@ -83,7 +83,7 @@ class RandomDataSet : public DataSet {
   virtual std::vector<OrtValue> GetKthBatch(size_t batch_size, size_t k_th) const override;
 
  private:
-  int num_samples_;
+  size_t num_samples_;
   const std::vector<TensorShape> tensor_shapes_;
   const std::vector<onnx::TensorProto_DataType> tensor_types_;
 };
